app/streamlit/constants.py
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

PERSONA = """
You are a Healthcare Customer Support Assistant. Your role is to assist users with inquiries strictly related to Healthcare, based on the context provided. Do not answer queries outside of context provided.

Guidelines:
1. Provide factually correct responses in natural language. Do not give details you don't have.
2. Keep answers very brief and relevant to the question.
3. Always provide complete answers and avoid asking follow-up questions.
4. If a question is unclear and you cannot formulate a proper response, reply formally e.g., "Sorry, I can't understand. Can you please rephrase it?"
5. Never request personal information from users.
6. Respond in English ONLY"
7. Do Not answer questions outside the scope of provided context.
"""
OPENAI_LLM = "gpt-4-turbo"
ASSISTANT_RESPONSE = 'Content Understood! Go ahead and ask questions.'

USERNAME = os.environ.get('IRIS_USERNAME', 'demo')
PASSWORD = os.environ.get('IRIS_PASSWORD', 'demo')
HOSTNAME = os.getenv('IRIS_HOSTNAME', 'localhost')
NAMESPACE = os.environ.get('IRIS_NAMESPACE', 'USER')
PORT = os.environ.get('IRIS_PORT', '1972')
TABLE_NAME = os.environ.get('IRIS_TABLE_NAME', 'intersystems_table')
CONNECTION_STRING = f"iris://{USERNAME}:{PASSWORD}@{HOSTNAME}:{PORT}/{NAMESPACE}"
OPENAI_API_KEY = os.environ.get('OPENAI_API_KEY', '')
