app/streamlit/requirements.txt
fastembed
ipykernel
langchain
langchain-iris
llama-iris
numpy
openai
pandas
pyodbc
PyPDF2
pypdf
python-dotenv
sentence-transformers
setuptools
sqlalchemy
sqlalchemy-iris
streamlit
testcontainers-iris
tiktoken
torch